fix: restore `BKTEvaluation.Reason` enum, move `ReasonType` back to `internal`
