+++ bucketeer/src/androidTest/kotlin/io/bucketeer/sdk/android/e2e/BKTClientVariationTest.kt
@@ -10,7 +10,6 @@ import io.bucketeer.sdk.android.BKTConfig
 import io.bucketeer.sdk.android.BKTEvaluation
 import io.bucketeer.sdk.android.BKTUser
 import io.bucketeer.sdk.android.BuildConfig
-import io.bucketeer.sdk.android.ReasonType
 import io.bucketeer.sdk.android.internal.Constants
 import io.bucketeer.sdk.android.internal.database.OpenHelperCallback
 import org.json.JSONObject
@@ -77,7 +76,7 @@ class BKTClientVariationTest {
         userId = USER_ID,
         variationId = "36a53a17-60b4-4a99-a54a-7fcbf21f7c8c",
         variationValue = "value-1",
-        reason = ReasonType.DEFAULT,
+        reason = BKTEvaluation.Reason.DEFAULT,
       ),
     )
   }
@@ -101,7 +100,7 @@ class BKTClientVariationTest {
         userId = USER_ID,
         variationId = "9b9a4396-d2ec-4eaf-aee6-ca0276881120",
         variationValue = "10",
-        reason = ReasonType.DEFAULT,
+        reason = BKTEvaluation.Reason.DEFAULT,
       ),
     )
   }
@@ -125,7 +124,7 @@ class BKTClientVariationTest {
         userId = USER_ID,
         variationId = "384bbcf0-0d1d-4e7a-b589-850f16f833b4",
         variationValue = "2.1",
-        reason = ReasonType.DEFAULT,
+        reason = BKTEvaluation.Reason.DEFAULT,
       ),
     )
   }
@@ -149,7 +148,7 @@ class BKTClientVariationTest {
         userId = USER_ID,
         variationId = "774fb34d-5b08-4305-9995-08cdac47aa0f",
         variationValue = "true",
-        reason = ReasonType.DEFAULT,
+        reason = BKTEvaluation.Reason.DEFAULT,
       ),
     )
   }
@@ -177,7 +176,7 @@ class BKTClientVariationTest {
         userId = USER_ID,
         variationId = "4499d1ca-411d-4ec6-9ae8-df51087e72bb",
         variationValue = """{ "key": "value-1" }""",
-        reason = ReasonType.DEFAULT,
+        reason = BKTEvaluation.Reason.DEFAULT,
       ),
     )
   }
