+++ bucketeer/src/main/kotlin/io/bucketeer/sdk/android/BKTClientImpl.kt
@@ -106,7 +106,7 @@ internal class BKTClientImpl(
       userId = raw.userId,
       variationId = raw.variationId,
       variationValue = raw.variationValue,
-      reason = ReasonType.from(raw.reason.type.name),
+      reason = BKTEvaluation.Reason.from(raw.reason.type.name),
     )
   }
