+++ bucketeer/src/main/kotlin/io/bucketeer/sdk/android/internal/model/Reason.kt
@@ -1,7 +1,6 @@
 package io.bucketeer.sdk.android.internal.model
 
 import com.squareup.moshi.JsonClass
-import io.bucketeer.sdk.android.ReasonType
 
 @JsonClass(generateAdapter = true)
 data class Reason(
