+++ bucketeer/src/main/kotlin/io/bucketeer/sdk/android/internal/model/ReasonType.kt
@@ -1,4 +1,4 @@
-package io.bucketeer.sdk.android
+package io.bucketeer.sdk.android.internal.model
 
 enum class ReasonType {
   TARGET,
