+++ bucketeer/src/test/kotlin/io/bucketeer/sdk/android/internal/evaluation/GetVariationValueTest.kt
@@ -3,9 +3,9 @@
 package io.bucketeer.sdk.android.internal.evaluation
 
 import com.google.common.truth.Truth.assertThat
-import io.bucketeer.sdk.android.ReasonType
 import io.bucketeer.sdk.android.internal.model.Evaluation
 import io.bucketeer.sdk.android.internal.model.Reason
+import io.bucketeer.sdk.android.internal.model.ReasonType
 import io.bucketeer.sdk.android.internal.model.Variation
 import org.json.JSONObject
 import org.junit.Test
